+++ sphinxcontrib/confluencebuilder/util.py
@@ -1,10 +1,7 @@
 # -*- coding: utf-8 -*-
 """
-    sphinxcontrib.confluencebuilder.util
-    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
     :copyright: Copyright 2018 by the contributors (see AUTHORS file).
-    :license: BSD, see LICENSE for details.
+    :license: BSD-2-Clause, see LICENSE for details.
 """
 
 from .std.confluence import API_REST_BIND_PATH
@@ -22,7 +19,7 @@ class ConfluenceUtil:
         """
         normalize a confluence base url
 
-        A Confluence base URL refers to the URL porition excluding the target
+        A Confluence base URL refers to the URL portion excluding the target
         API bind point. This method attempts to handle a series of user-provided
         URL values and attempt to determine the proper base URL to use.
         """